feat(google-maps-builder):feat - Descriptions with Hint.css tooltips replaced

+++ includes/libraries/metabox/includes/CMB2.php
@@ -390,10 +390,20 @@ class CMB2 extends CMB2_Base {
 		$field_group->index = 0;
 
 		$field_group->peform_param_callback( 'before_group' );
+		$render_existing_marker_tooltip = $field_group->args( 'render_existing_marker_tooltip' );
 
 		echo '<div class="cmb-row cmb-repeat-group-wrap ', $field_group->row_classes(), '" data-fieldtype="group"><div class="cmb-td"><div id="', $field_group->id(), '_repeat" class="cmb-nested cmb-field-list cmb-repeatable-group', $sortable, $repeat_class, '" style="width:100%;">';
 
-		if ( $desc || $label ) {
+		if ( $render_existing_marker_tooltip ) {
+			echo sprintf(
+				'<div class="maps-marker-label"><h2 class="cmb-group-name">%1$s</h2><span class="hint--top hint--top-multiline" aria-label="%2$s"><span class="dashicons gmb-tooltip-icon"></span></span>
+				<a href="#" class="button button-small toggle-repeater-groups">' . __( "Toggle Marker Groups", "google-maps-builder" ) . "</a></div>",
+				esc_html__( "Existing Markers", "google-maps-builder" ),
+				esc_html__( 'Map marker data is contained within the repeatable fields below. You may add or update marker data here or directly on the map. ', 'google-maps-builder' )
+			);
+
+		}
+		if ( ( $desc || $label ) && empty( $render_existing_marker_tooltip ) ) {
 			$class = $desc ? ' cmb-group-description' : '';
 			echo '<div class="cmb-row', $class, '"><div class="cmb-th">';
 			if ( $label ) {
@@ -450,21 +460,11 @@ class CMB2 extends CMB2_Base {
 		$desc      = $field_group->args( 'description' );
 		$label     = $field_group->args( 'name' );
 		$group_val = (array) $field_group->value();
-		$render_existing_marker_tooltip     = $field_group->args( 'render_existing_marker_tooltip' );
+
 
 		echo '<div class="cmb-row cmb-repeat-group-wrap ', esc_attr( $field_group->row_classes() ), '" data-fieldtype="group"><div class="cmb-td"><div data-groupid="', esc_attr( $field_group->id() ), '" id="', esc_attr( $field_group->id() ), '_repeat" ', $this->group_wrap_attributes( $field_group ), '>';
-		if ($render_existing_marker_tooltip) {
 
-			echo sprintf(
-				'<div class="maps-marker-label"><h2 class="cmb-group-name">%1$s</h2><span class="hint--top hint--top-multiline" aria-label="%2$s"><span class="dashicons gmb-tooltip-icon"></span></span>
-				<a href="#" class="button button-small toggle-repeater-groups">' . __( "Toggle Marker Groups", "google-maps-builder" ) . "</a></div>",
-				esc_html__( "Existing Markers", "google-maps-builder" ),
-				esc_html__( 'Map marker data is contained within the repeatable fields below. You may add or update marker data here or directly on the map. ', 'google-maps-builder' )
-			);
-
-			//echo '<h2 class="cmb-group-name">', $render_existing_marker_tooltip, '</h2>';
-		}
-		if ( ( $desc || $label ) && empty($render_existing_marker_tooltip) ) {
+		if ( ( $desc || $label ) ) {
 			$class = $desc ? ' cmb-group-description' : '';
 			echo '<div class="cmb-row', $class, '"><div class="cmb-th">';
 			if ( $label ) {
